requirements.txt
aioredis==1.3.1
asgiref==3.3.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
channels==3.0.0
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
cymem==2.0.4
daphne==3.0.0
Django==3.1.3
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
django-resized==0.3.11
django-taggit==1.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
msgpack==1.0.0
murmurhash==1.0.4
numpy==1.19.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
pytz==2020.4
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
spacy==2.2.2
sqlparse==0.4.1
srsly==1.0.2
thinc==7.3.1
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.51.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.11
wasabi==0.8.0
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.1.2
